Split router module into multiple files
